configure.ac: fix check for egt headers
The AC_CHECK_HEADERS macro that is testing for certain egt headers to be
installed to determine if certain egt functionality is present assumes
that egt is installed with the normal system headers.  However,
especially during development it is possible that egt is not installed
in /usr/include but kept in a local directory.  Take the CFLAGS returned
from the PKG_CHECK_MODULES macro to determine where those egt headers
are installed.

Signed-off-by: Dan Sneddon <daniel.sneddon@microchip.com>
[change variable names]
Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>