AngryMeatballs: abstract the coordinate of the ground in the EGT window
Add MainWindow::world2egt() and MainWindow::egt2world() methods to
abstract the coordinate of the ground in the EGT window.

This way, we hide the ground coordinate from the Ball class.

Both world2egt() and egt2world() methods rely on MainWindow::ground() to
get the ground coordinate.

Finally, this ground() method computes the coordinate of the ground, by
scaling the screen height with a new 'm_ground_ratio' member.

This way, the ground is correctly positioned even if the image
background is streched to fit the actual screen size.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>